change app to recalculate crosscut data everytime the grooves button is clicked
Workflow:

1. Click upload and select x3p file: scan gets loaded and displayed
2. Click crosscut button: "optimal" crosscut location (y-value) is found and displayed on scan
3. Move slider or click slider buttons: crosscut location is adjusted
4. Click grooves button: crosscut profile is taken at the current crosscut location. Crosscut and grooves are plotted.

Previously, if you didn't like the crosscut profile on step 4, you could go back to step 3 and adjust the crosscut location. But when you click the grooves button in step 4 a second time, the crosscut profile wasn't updated. This is fixed so that now the crosscut profile is updated every time you click the grooves button, not just the first time.