add groove sliders to land scan display
Adjusting the location of the grooves from the crosscut profile plot is difficult. Looking at the land scans, it is easy to see that there is almost always missing data in the deepest part of the groove. But it is hard to tell on the profile plot where data is missing. I moved the grooves sliders to the land scan display to so the user can reference the scan to pick the groove locations instead of relying on the crosscut profile.