change grooves tab to profile tab
Now that the groove locations are displayed and adjusted on the land scan along with the crosscut, I changed the grooves tab to preview the profile. The profile is the surface height of the land at the current crosscut location. The left and right groove locations are plotted on the profile, but cannot be changed on the profile tab. Because of interpolation, the profile plot makes it look like there is data where there is not. Forcing users to place the groove locations on the land scan keeps them from mistakenly including interpolated shoulder data in the profile.