change grooves slider limits
Previously, the slider limits for the grooves were set at 0 and the full width of the scan. But this makes it hard to make small changes with the sliders. Initially, I changed the sliders' range to 1/2 the width of the scan, but because there is a ton of white space on the right side of the scan, 1/2 the width is close to the right side of the land. Instead, I changed the sliders' range to be [0, width/3] for the left slider and [width/3, width] for the right slider.